Posible final

+++ AD/Cuarto_Proyecto_AD/ProyectoFinal/Reservas/src/main/java/com/proyecto/reservas/dto/ActualizarHabitacionDTO.java
@@ -3,13 +3,15 @@ package com.proyecto.reservas.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class ActualizarHabitacionDTO {
     private int id;
     private int numeroHabitacion;
     private String tipo;
-    private double precio;
+    private BigDecimal precio;
     private int idHotel;
     private Boolean disponible;
 }
